requirements.txt
streamlit
pydantic-ai
python-dotenv
requests
pandas
pytest
pytest-asyncio
black
flake8
mypy
msal
